refactor(spi): move ContractNegotiationCommand
